+++ contracts/types/CommunityTypes.sol
@@ -89,9 +89,9 @@ struct Invitation {
  * `false` otherwise.
  */
 struct UserTypeSetting {
-  uint256 proportionalityOnRegister;
+  uint8 proportionalityOnRegister;
   bool directProportionalityRegistration;
   bool needInvitationOnRegister;
-  uint256 invitationDelayBlocks;
+  uint32 invitationDelayBlocks;
   bool isVoter;
 }
